@@ -22,7 +22,9 @@ rootProject.allprojects {
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 28
+    namespace 'pt.tribeiro.flutter_plugin_pdf_viewer'
+
+    compileSdkVersion 34
 
     defaultConfig {
         minSdkVersion 21
